协议修改

+++ meetingManager/src/main/scala/videomeeting/meetingManager/core/MeetingManager.scala
@@ -34,7 +34,7 @@ object MeetingManager {
 
   case class GetMeetingList(replyTo: ActorRef[MeetingListRsp]) extends Command
 
-  case class SearchMeeting(userId:  Option[Int], roomId: Int, replyTo:ActorRef[SearchMeetingRsp]) extends Command
+  case class SearchMeeting(userId:  Option[Long], roomId: Long, replyTo:ActorRef[SearchMeetingRsp]) extends Command
 
   case class UserInfoChange(userId:Int,temporary:Boolean) extends Command
 
@@ -203,7 +203,7 @@ object MeetingManager {
     }.unsafeUpcast[MeetingActor.Command]
   }
 
-  private def getMeetingActorOpt(meetingId:Int, ctx: ActorContext[Command]) = {
+  private def getMeetingActorOpt(meetingId:Long, ctx: ActorContext[Command]) = {
     val childrenName = s"roomActor-${meetingId}"
 //    log.debug(s"${ctx.self.path} the child = ${ctx.children},get the roomActor opt = ${ctx.child(childrenName).map(_.unsafeUpcast[RoomActor.Command])}")
     ctx.child(childrenName).map(_.unsafeUpcast[MeetingActor.Command])
